pyproject.toml
[tool.poetry]
name = "car-wash-backend"
version = "0.1.0"
description = ""
authors = ["PavelPrist <umcperovo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
pre-commit = "^3.5.0"
drf-spectacular = "^0.26.5"
djangorestframework = "^3.14.0"
django-filter = "^23.4"
python-dotenv = "^1.0.0"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.2.0"}
gunicorn = "^21.2.0"
taskipy = "^1.12.2"
psycopg2-binary = "^2.9.9"
django-multiselectfield = "^0.1.12"
geopy = "^2.4.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
start = "gunicorn car_wash_backend.wsgi:application --bind"
migrate = "python manage.py migrate"
createsuperuser = "python manage.py createsuperuser"
collectstatic = "python manage.py collectstatic --no-input"
